moved Playback & Recording into Amp (they are functional units now). Deleted StreamCollector (will now revive timeline)
