--- scalaudio-core/src/main/scala/scalaudio/core/engine/io/Recording.scala
@@ -1,31 +1,0 @@
-package scalaudio.core.engine.io
-
-import java.io.File
-
-import com.jsyn.util.WaveFileWriter
-
-import scalaudio.core.AudioContext
-import scalaudio.core.engine.Interleaver
-import scalaudio.core.types._
-
-/**
-  * Created by johnmcgill on 1/6/16.
-  */
-case class Recording(filename: String)(implicit audioContext: AudioContext) extends OutputEngine {
-  val waveFile: File = new File(filename + ".wav")
-  val writer = new WaveFileWriter(waveFile)
-  writer.setFrameRate(audioContext.config.samplingRate)
-  writer.setSamplesPerFrame(audioContext.config.nOutChannels)
-  writer.setBitsPerSample(16)
-
-  override def start(): Unit = println(s"Writing to WAV file ${waveFile.getAbsolutePath}")
-
-  override def stop(): Unit = writer.close()
-
-  override def handleBuffers(output: Either[AudioSignal, MultichannelAudio]) = record(output)
-
-  def record(output: Either[AudioSignal, MultichannelAudio]) = output match {
-    case Left(audioSignal) => writer.write(audioSignal)
-    case Right(multichannelAudio) => writer.write(Interleaver.interleave(multichannelAudio))
-  }
-}
